+++ src/components/message.js
@@ -4,6 +4,7 @@ import './message-actions.js';
 import MessageVersionsModal from '../modals/message-versions.js';
 import dayjs from 'dayjs';
 import filesize from "filesize";
+import tpl_chat_message from '../templates/chat_message.js';
 import tpl_spinner from '../templates/spinner.js';
 import { CustomElement } from './element.js';
 import { __ } from '@converse/headless/i18n';
@@ -18,23 +19,21 @@ const i18n_edited = __('This message has been edited');
 const i18n_show = __('Show more');
 const i18n_show_less = __('Show less');
 const i18n_uploading = __('Uploading file:');
-const i18n_new_messages = __('New messages');
 
 
 class Message extends CustomElement {
 
     static get properties () {
         return {
-            allow_retry: { type: Boolean },
             chatview: { type: Object},
             correcting: { type: Boolean },
             editable: { type: Boolean },
+            edited: { type: String },
             error: { type: String },
             error_text: { type: String },
             from: { type: String },
             has_mentions: { type: Boolean },
             hats: { type: Array },
-            edited: { type: String },
             is_delayed: { type: Boolean },
             is_encrypted: { type: Boolean },
             is_first_unread: { type: Boolean },
@@ -55,6 +54,7 @@ class Message extends CustomElement {
             reason: { type: String },
             received: { type: String },
             retractable: { type: Boolean },
+            retry_event_id: { type: String },
             sender: { type: String },
             show_spinner: { type: Boolean },
             spoiler_hint: { type: String },
@@ -105,7 +105,7 @@ class Message extends CustomElement {
                 </div>
                 ${ this.reason ? html`<q class="reason">${this.reason}</q>` : `` }
                 ${ this.error_text ? html`<q class="reason">${this.error_text}</q>` : `` }
-                ${ this.allow_retry ? html`<a class="retry" @click=${this.onRetryClicked}>${i18n_retry}</a>` : '' }
+                ${ this.retry_event_id ? html`<a class="retry" @click=${this.onRetryClicked}>${i18n_retry}</a>` : '' }
             </div>
         `;
     }
@@ -124,42 +124,7 @@ class Message extends CustomElement {
     }
 
     renderChatMessage () {
-        const is_groupchat_message = (this.message_type === 'groupchat');
-        return html`
-            ${ this.is_first_unread ? html`<div class="message date-separator"><hr class="separator"><span class="separator-text">${ i18n_new_messages }</span></div>` : '' }
-            <div class="message chat-msg ${ this.getExtraMessageClasses() }"
-                 data-isodate="${this.time}"
-                 data-msgid="${this.msgid}"
-                 data-from="${this.from}"
-                 data-encrypted="${this.is_encrypted}">
-
-                ${ this.shouldShowAvatar() ? renderAvatar(this.getAvatarData()) : '' }
-                <div class="chat-msg__content chat-msg__content--${this.sender} ${this.is_me_message ? 'chat-msg__content--action' : ''}">
-                    <span class="chat-msg__heading">
-                        ${ (this.is_me_message) ? html`
-                            <time timestamp="${this.time}" class="chat-msg__time">${this.pretty_time}</time>
-                            ${this.hats.map(hat => html`<span class="badge badge-secondary">${hat}</span>`)}
-                        ` : '' }
-                        <span class="chat-msg__author">${ this.is_me_message ? '**' : ''}${this.username}</span>
-                        ${ !this.is_me_message ? this.renderAvatarByline() : '' }
-                        ${ this.is_encrypted ? html`<span class="fa fa-lock"></span>` : '' }
-                    </span>
-                    <div class="chat-msg__body chat-msg__body--${this.message_type} ${this.received ? 'chat-msg__body--received' : '' } ${this.is_delayed ? 'chat-msg__body--delayed' : '' }">
-                        <div class="chat-msg__message">
-                            ${ this.is_retracted ? this.renderRetraction() : this.renderMessageText() }
-                        </div>
-                        ${ (this.received && !this.is_me_message && !is_groupchat_message) ? html`<span class="fa fa-check chat-msg__receipt"></span>` : '' }
-                        ${ (this.edited) ? html`<i title="${ i18n_edited }" class="fa fa-edit chat-msg__edit-modal" @click=${this.showMessageVersionsModal}></i>` : '' }
-                        <converse-message-actions
-                            .chatview=${this.chatview}
-                            .model=${this.model}
-                            ?correcting="${this.correcting}"
-                            ?editable="${this.editable}"
-                            ?is_retracted="${this.is_retracted}"
-                            message_type="${this.message_type}"></converse-message-actions>
-                    </div>
-                </div>
-            </div>`;
+        return tpl_chat_message(this);
     }
 
     shouldShowAvatar () {
@@ -180,7 +145,7 @@ class Message extends CustomElement {
 
     async onRetryClicked () {
         this.show_spinner = true;
-        await this.model.error.retry();
+        await api.trigger(this.retry_event_id, {'synchronous': true});
         this.model.destroy();
         this.parentElement.removeChild(this);
     }
@@ -248,6 +213,7 @@ class Message extends CustomElement {
     }
 
     renderMessageText () {
+        const is_groupchat_message = (this.message_type === 'groupchat');
         const tpl_spoiler_hint = html`
             <div class="chat-msg__spoiler-hint">
                 <span class="spoiler-hint">${this.spoiler_hint}</span>
@@ -257,16 +223,22 @@ class Message extends CustomElement {
                 </a>
             </div>
         `;
+        const spoiler_classes = this.is_spoiler ? `spoiler ${this.is_spoiler_visible ? '' : 'collapsed'}` : '';
         return html`
             ${ this.is_spoiler ? tpl_spoiler_hint : '' }
             ${ this.subject ? html`<div class="chat-msg__subject">${this.subject}</div>` : '' }
-            <converse-chat-message-body
-                .model="${this.model}"
-                ?is_me_message="${this.is_me_message}"
-                ?is_only_emojis="${this.is_only_emojis}"
-                ?is_spoiler="${this.is_spoiler}"
-                ?is_spoiler_visible="${this.is_spoiler_visible}"
-                text="${this.model.getMessageText()}"></converse-chat-message-body>
+            <span>
+                <converse-chat-message-body
+                    class="chat-msg__text ${this.is_only_emojis ? 'chat-msg__text--larger' : ''} ${spoiler_classes}"
+                    .model="${this.model}"
+                    ?is_me_message="${this.is_me_message}"
+                    ?is_only_emojis="${this.is_only_emojis}"
+                    ?is_spoiler="${this.is_spoiler}"
+                    ?is_spoiler_visible="${this.is_spoiler_visible}"
+                    text="${this.model.getMessageText()}"></converse-chat-message-body>
+                ${ (this.received && !this.is_me_message && !is_groupchat_message) ? html`<span class="fa fa-check chat-msg__receipt"></span>` : '' }
+                ${ (this.edited) ? html`<i title="${ i18n_edited }" class="fa fa-edit chat-msg__edit-modal" @click=${this.showMessageVersionsModal}></i>` : '' }
+            </span>
             ${ this.oob_url ? html`<div class="chat-msg__media">${u.getOOBURLMarkup(_converse, this.oob_url)}</div>` : '' }
             <div class="chat-msg__error">${ this.error_text || this.error }</div>
         `;