+++ src/converse-roomslist.js
@@ -10,8 +10,8 @@ import "@converse/headless/converse-muc";
 import RoomDetailsModal from 'modals/muc-details.js';
 import { _converse, api, converse } from "@converse/headless/converse-core";
 import tpl_rooms_list from "templates/rooms_list.js";
-import { Model } from 'skeletor.js/src/model.js';
-import { View } from 'skeletor.js/src/view.js';
+import { Model } from '@converse/skeletor/src/model.js';
+import { View } from '@converse/skeletor/src/view.js';
 import { __ } from '@converse/headless/i18n';