+++ src/headless/utils/core.js
@@ -4,8 +4,8 @@
  * @description This is the core utilities module.
  */
 import * as strophe from 'strophe.js/src/core';
-import { Model } from 'skeletor.js/src/model.js';
-import { compact, last, isElement, isObject, isString } from "lodash";
+import { Model } from '@converse/skeletor/src/model.js';
+import { compact, last, isElement, isObject, isString } from "lodash-es";
 import log from "@converse/headless/log";
 import sizzle from "sizzle";